+++ apps/dashboard/app/dashboard/Footer.tsx
@@ -1,11 +1,15 @@
+"use client"
+
 import { NextPage } from "next";
 import { ReactNode } from "react";
 import Image from "next/image";
 import { Avatar, Dropdown, Navbar, Footer } from "flowbite-react";
-import { authOptions } from "../pages/api/auth/[...nextauth]";
+import { authOptions } from "../../src/pages/api/auth/[...nextauth]";
 import { useSession, signIn, signOut } from "next-auth/react";
 
-const MainFooter: NextPage = () => {
+
+
+export default function MainFooter() {
   return (
     <div>
       <Footer container={true}>
@@ -20,5 +24,3 @@ const MainFooter: NextPage = () => {
     </div>
   );
 };
-
-export default MainFooter;
